backend/requirements.txt
celery
redis
pillow
python-dotenv
boto3
